+++ server/index.js
@@ -3,11 +3,14 @@ const req = require("express/lib/request");
 const res = require("express/lib/response");
 const User = require("./models/user");
 const Joi = require("joi");
+const loginRoute = require("./router/login");
+
+
 
 const app = express();
 require("dotenv").config();
 const mongoCred = process.env.MONGO_CRED;
-const dbAccess = `mongodb+srv://test:${mongoCred}@kuchtohai.cuvoy.mongodb.net/kuchtohai?retryWrites=true&w=majority`;
+const dbAccess = `mongodb+srv://testdepix:testdepix@cluster0.pqtoa.mongodb.net/testdepix?retryWrites=true&w=majority`;
 const mongoose = require("mongoose");
 const { get } = require("express/lib/response");
 const jIO = require("jio");
@@ -30,6 +33,8 @@ app.get("/users", (req, res) => {
     res.send(result);
   });
 });
+app.use(loginRoute)
+
 
 //Getting the Perticular User information stored in Server
 app.get("/users/:token", (req, res) => {
@@ -51,29 +56,6 @@ app.get("/users/:token", (req, res) => {
   //   res.status(404).send({ stattus: 404, result: null });
 });
 
-//Creating the User
-app.post("/users", (req, res) => {
-  console.log("requesting Post  method...........!!!!");
-  const token = req.body.token;
-  const result = findUserByToken(token);
-  console.log(result);
-
-  if (result.userExist) {
-    return res
-      .status(302)
-      .send({ userExist: true, result: null, userCreated: false, status: 302 });
-  }
-
-  const user = new User({ token: token });
-  user.save().then((_result) => {
-    res.status(201).send({
-      userExist: false,
-      result: _result,
-      userCreated: true,
-      staus: 201,
-    });
-  });
-});
 
 //Adding nft to User
 app.post("/nft", (req, res) => {
@@ -104,44 +86,3 @@ app.post("/nft", (req, res) => {
     status: 201,
   });
 });
-
-function validate(body) {
-  const schema = {
-    token: Joi.string().required(),
-    nft: Joi.string().required(),
-  };
-  return Joi.validate(body, schema);
-}
-
-function findUserByToken(token) {
-  let result;
-  User.find({ token: token }, (error, data) => {
-    if (error) result = null;
-    if (data) result = data;
-  });
-  console.log(result);
-  if (!result) {
-    return {
-      status: 404,
-      userExist: false,
-      result: null,
-    };
-  }
-
-  return {
-    status: 302,
-    userExist: true,
-    result: result,
-  };
-}
-
-async function addNftToId(_id, nft) {
-  return User.findByIdAndUpdate(
-    { _id },
-    {
-      $addToSet: {
-        nfts: nft,
-      },
-    }
-  );
-}
